requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
bitarray==1.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
EasyProcess==0.2.7
idna==2.8
Logbook==1.4.3
lxml==4.3.4
pycparser==2.19
PyExecJS==1.5.1
PySocks==1.7.0
python-dateutil==2.8.0
PyVirtualDisplay==0.2.4
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
